+++ Vortex2D/Engine/Pressure.h
@@ -16,7 +16,7 @@ class Pressure
 public:
     Pressure(const Renderer::Device& device,
              float dt,
-             const glm::vec2& size,
+             const glm::ivec2& size,
              LinearSolver& solver,
              Renderer::Texture& velocity,
              Renderer::Texture& solidPhi,
